pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "arc-cli"
version = "0.1.0"
description = "Agent Router CLI -- lightweight agent dispatch and scheduling"
requires-python = ">=3.11"
dependencies = [
    "typer>=0.12",
    "httpx>=0.27",
    "apscheduler>=3.10,<4",
    "discord.py>=2.3",
    "pyyaml>=6.0",
]

[project.scripts]
arc = "arc.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/arc"]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-httpx>=0.30",
    "pytest-cov>=5.0",
    "ruff>=0.4",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]
